pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "geopardie"
description = "Steve Harvey Simulator Discord Bot"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "LGPLv3" }
authors = [{ name = "Mingde Yin", email = "mdsuper@hotmail.com" }]
maintainers = [{ name = "Mingde Yin", email = "mdsuper@hotmail.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
]
dependencies = ["naff>=2.1.0", "python-dotenv>=0.21.0", "orjson"]
version = "0.1.0"

[project.urls]
homepage = "https://github.com/itchono/geopardie"

[tool.setuptools.packages.find]
include = ["geopardie"]
where = ["src"]

[tool.autopep8]
exclude = "venv"
max_line_length = 80
in-place = true
recursive = true
aggressive = 3
